Move unit compat test to existing suite
